Consolidated rabbit test
One test can handle testing publish, receiving, and delay. It'd be a lot
of copy paste for no meaningful gain if I were to separate them. Doing
so...

+++ Common/src/main/java/WebsiteMonitor/RabbitPublisher.java
@@ -39,7 +39,7 @@ public class RabbitPublisher {
         rabbitQueueName = queueName;
     }
 
-    public void EnqueueTask(Task task, int taskDelayMs) throws IOException {
+    public void EnqueueTask(Task task, long taskDelayMs) throws IOException {
         Gson gson = new Gson();
         String taskAsJson = gson.toJson(task);
